Agents_Systems/chat_with_gmail/requirements.txt
streamlit
embedchain[gmail]
python-dotenv
sentence-transformers
pydantic-settings
chromadb<0.4.22
langchain-community
langchain-text-splitters